recommendation

+++ src/main/java/org/example/lv_backend/dto/request/chapter/ChapterUpdateRequest.java
@@ -15,7 +15,7 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 public class ChapterUpdateRequest {
     @NotNull(message = "CHAPTER_NUMBER_NULL")
-    private Long chapterNumber;
+    private Integer chapterNumber;
 
     @NotBlank(message = "TITLE_BLANK")
     private String title;
